+++ Changelog.md
@@ -17,6 +17,7 @@
 - [Fix invar block recipe in alloy smelter](https://github.com/TheStaticVoid/StaTech-Industry/issues/337)
 - [Fix spectrum saplings not being a chance output](https://github.com/TheStaticVoid/StaTech-Industry/issues/340)
 - [Add TrollDespair to Emojiful](https://github.com/TheStaticVoid/StaTech-Industry/issues/347)
+- [Add Marble macerator recipe](https://github.com/TheStaticVoid/StaTech-Industry/issues/344)
 
 ## First-time Contributors
 - redcatone